Rectangle Drawing Gui Fiiles All Located Here
Project I made with a partner for my Programming Method: Java class (which is essentially the second part of Intro to Object Oriented Programming: Java)  in my Fall 2019 semester. Using JavaFx we made a GUI component where you can click to draw either a rectangle or a circle. You are also given the option to fill in the shape and give it thick or thin borders.